Tasks working well

+++ src/app/Presentation/UserController.py
@@ -3,6 +3,7 @@ from flask_restx import Resource, Namespace
 from injector import inject
 from app.Presentation.Resources.UserResource import user_resource
 from app.Security.TokenTools import token_required
+from flask import g
 
 
 api = Namespace('users', description='User related operations')
@@ -17,26 +18,26 @@ class UserList(Resource):
     def __init__(self,user_service : UserService, **kwargs):
         self.user_service = user_service
         super().__init__(**kwargs)
-
-    api.doc('Lists_users')
-    #@api.expect(parser_user_lists)
-    @api.marshal_list_with(ReadUser)
-    @token_required
-    def get(self):
-        """List all users"""
-        users = self.user_service.get_all()
-        return users, 200
+    # FIXME: Implement roles per user so we can use this
+    # api.doc('Lists_users')
+    # #@api.expect(parser_user_lists)
+    # @api.marshal_list_with(ReadUser)
+    # @token_required
+    # def get(self):
+    #     """List all users"""
+    #     users = self.user_service.get_all()
+    #     return users, 200
     
-    @api.doc('Create_user')
-    @api.expect(CreateUserDto)
-    @api.marshal_with(ReadUser)
-    @token_required
-    def post(self):
-        """Create a new user"""
-        data = api.payload.copy() #Is what the post brings
-        print(data)
-        user = self.user_service.create(**data)
-        return user, 201
+    # @api.doc('Create_user')
+    # @api.expect(CreateUserDto)
+    # @api.marshal_with(ReadUser)
+    # @token_required
+    # def post(self):
+    #     """Create a new user"""
+    #     data = api.payload.copy() #Is what the post brings
+    #     print(data)
+    #     user = self.user_service.create(**data)
+    #     return user, 201
     
 
 @api.route('/<int:id>')
@@ -66,7 +67,10 @@ class User(Resource):
     def put(self, id):
         """Update a user given its identifier"""
         data = api.payload
-        user = self.user_service.update(id, data)
+        if id == g.user_id:
+            user = self.user_service.update(id, data)
+        else:
+            api.abort(404)
         if not user:
             api.abort(404)
         else:
@@ -77,7 +81,10 @@ class User(Resource):
     @api.response(204, 'User deleted')
     def delete(self, id):
         """Delete a user"""
-        user = self.user_service.delete(id)
+        if id == g.user_id:
+            user = self.user_service.delete(id)
+        else:
+            api.abort(404)
         if not user:
             api.abort(404)
         else:
